Improved develop page for plugin developers
I removed many links from the develop menu that are only important
for core developers or pro developers to a section named Core development
which can be reached from the develop menu (very bottom of the menu) or
from the develop-introduction page. We could maybe also name it
"Piwik in depth" or similar instead of "Core development" in case someone
is interested in those details.

I splitted some pages already in 2, eg tests-travis for plugin developers
and tests-travis for core.

Getting started is now only one page. The second part is now under Plugin
basics as "Custom Report". The third getting started page is completely
removed as it is explained in the other guides anyway.

The menu contained many links before that are not important for the majority
as we are only a few core developers and some of us do not even really use
developer.piwik.org as we are aware of this stuff. It should be still reachable.
That's why there are links to the core section and some guides might link to
the extended guides in core directly.