Remove the example on how to call the API through PHP (#246)
This is not needed and HTTP should always be recommended. There is not really a performance or resources improvement over going through HTTP. Also the example doesn't work and would need to be updated regularly. None of those methods mentioned there are API and it is only a matter of time until it breaks (it did in the past eg when we added environment etc).