Remove set calling setSiteId after creating the tracker (#362)
I'd recommend to not mention this anymore as ideally since Matomo 3.14.0 this should be set directly in the tracker. This guarantees that when the tracker event `TrackerAdded` is triggered, that the tracker URL and tracker idSite is configured. Heatmaps and Session Recording listens to this event and tries to fetch the `configs.php` based on this event. However, it can only do this if idSite and URL is configured already. There's no event when idSite is set or anything similar. There can be otherwise edge cases when creating tracker instances manually where the configs.php is not requested when configuring the idSite later.