Update profiling-code.md with SQL_NO_CACHE warning
The documentation recommends that people use `SQL_NO_CACHE` when running different performance queries and profiling. However, we have to be careful to warn of a few SQL parser limitations that will cause the desired effect to be ignored.
 
I've also updated a few hints to what versions this applies to. SQL_NO_CACHE has been deprecated at MySQL 5.7.20, so I've noted that.